App Complete!
fully functioning 'movie-fight' app where you can compare 2 different movies!

the BoxOffice value displays 'N/A' because there is no actual value provided from the database; there is a value for 'BoxOffice' but there is no value assigned to the variable. I left it in the app nonetheless for my own personal reference. If there are movies with a value assigned I would like that comparison to be made. 
For professional / final draft purposes maybe I would remove it.